I Love Zach, but not his code. I fixed it now time to do the assignment
Thanks Jeff
