+++ bin/.gitignore
@@ -2,3 +2,4 @@
 /Account.class
 /module-info.class
 /ATMTester.class
+/atm_package/
