--- src/ATM.java
@@ -1,20 +1,0 @@
-import java.util.*;
-public class ATM {
-	private ArrayList<Account> accounts;
-	public double deposit (String account, double amount)
-	{
-		
-		if(!accounts.contains(new Account(account)))
-		{
-			Account newAcc = new Account(account);
-			accounts.add(newAcc);
-			newAcc.deposit(amount);
-			return newAcc.getBalance(); 
-		}
-		else
-		{
-			accounts.get(accounts.indexOf(new Account(account))).deposit(amount);
-		}
-		
-	}
-}
