+++ src/atm_package/ATMTester.java
@@ -1,9 +1,9 @@
-
+package atm_package;
 public class ATMTester {
 	public static void main (String [] args)
 	{
 		ATM jeremiah = new ATM (); 
-		System.out.print(jeremiah.deposit("zach", 60)); 
+		System.out.println(jeremiah.deposit("zach", 60)); 
 		System.out.print(jeremiah.deposit("zach", 60)); 
 	}
 }
